- Move functions that are used both by Verifier and Prover module up a hierachy.
- Remove check on xy consistency
